add support for "known blocks" implemented as deleted block indexes; streamline block index creation and deletion
A known block is a block that is not part of the block tree, but is stored in the in-mem block store and can be (re-)added to the block tree by restore()'ing it and its ancestors.
Removing a block from the block tree converts it into a known block.

This should reduce the number of times block hashes are calculated and allow algorithms to operate on multiple chains at once and attach metadata to non-best chain blocks.